Move user list page component (#178)
