+++ frontend/app/routes/home.users._index/components/UsersListComponent.tsx
@@ -2,11 +2,11 @@ import { Stack } from '@mantine/core';
 import { SerializeFrom } from '@remix-run/cloudflare';
 import { getUsersResponse } from 'client/client';
 import { FaUsers } from 'react-icons/fa';
+import BreadCrumbsComponent from '~/components/common/breadcrumbs/BreadCrumbsComponent';
+import ErrorComponent from '~/components/common/error/ErrorComponent';
+import ContentsHeader from '~/components/common/pagination/ContentsHeader';
+import PaginationComponent from '~/components/common/pagination/PaginationComponent';
 import type { PaginationProps } from '~/types/pagination';
-import BreadCrumbsComponent from '../common/breadcrumbs/BreadCrumbsComponent';
-import ErrorComponent from '../common/error/ErrorComponent';
-import ContentsHeader from '../common/pagination/ContentsHeader';
-import PaginationComponent from '../common/pagination/PaginationComponent';
 import UserCreateButton from './UserCreateButton';
 import UsersListTable from './UsersListTable';
 import UsersListTitle from './UsersListTitle';
